Auto stash before merge of "main" and "origin/main"
