+++ src/components/NewsletterForm/Newsletter.Model.tsx
@@ -1,23 +1,5 @@
 import { Tag } from '@models/note'
 
-export enum LoadingState {
-  INIT = `INIT`,
-  REQUESTING = `REQUESTING`,
-  REQUESTING_MORE = `REQUESTING_MORE`,
-  PROCESSING = `PROCESSING`,
-  RESOLVED = `RESOLVED`,
-  FAILED = `FAILED`,
-  REJECTED = `REJECTED`,
-  STOPPING = `STOPPING`,
-  STOPPED = `STOPPED`,
-}
-
-export interface ErrorState {
-  errorMsg: string
-}
-
-export type RequestState = LoadingState | ErrorState
-
 export interface NewsletterBaseProps {
   tags?: Array<Tag>
 }
